requirements_test.txt
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-homeassistant-custom-component==0.7.0
requests>=2.32.3
python-dotenv>=1.0.1
